+++ connection.go
@@ -268,23 +268,7 @@ func New(ctx context.Context, opts ...Option) (_ Connection, err error) {
 
 	c.coordination = lazy.Coordination(c.db, c.coordinationOptions)
 
-	c.ratelimiter = lazy.Ratelimiter(
-		c.db,
-		append(
-			// prepend operation timeout and cancelAfter options before custom ratelimiterOptions
-			// If custom ratelimiterOptions contains operation timeout or cancelAfter options
-			// - will apply custom values
-			[]ratelimiterConfig.Option{
-				ratelimiterConfig.WithOperationTimeout(
-					c.config.OperationTimeout(),
-				),
-				ratelimiterConfig.WithOperationCancelAfter(
-					c.config.OperationCancelAfter(),
-				),
-			},
-			c.ratelimiterOptions...,
-		),
-	)
+	c.ratelimiter = lazy.Ratelimiter(c.db, c.ratelimiterOptions)
 
 	return c, nil
 }