+++ test/ratelimiter_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/config"
 	cfg "github.com/ydb-platform/ydb-go-sdk/v3/coordination"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
-	public "github.com/ydb-platform/ydb-go-sdk/v3/ratelimiter"
+	"github.com/ydb-platform/ydb-go-sdk/v3/ratelimiter"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
@@ -81,9 +81,9 @@ func TestRatelimiter(t *testing.T) {
 		}
 	}()
 	// create resource
-	err = db.Ratelimiter().CreateResource(ctx, testCoordinationNodePath, public.Resource{
+	err = db.Ratelimiter().CreateResource(ctx, testCoordinationNodePath, ratelimiter.Resource{
 		ResourcePath: testResource,
-		HierarchicalDrr: public.HierarchicalDrrSettings{
+		HierarchicalDrr: ratelimiter.HierarchicalDrrSettings{
 			MaxUnitsPerSecond:       1,
 			MaxBurstSizeCoefficient: 2,
 		},
@@ -110,9 +110,9 @@ func TestRatelimiter(t *testing.T) {
 		t.Fatal("Resource invalid")
 	}
 	// alter resource
-	err = db.Ratelimiter().AlterResource(ctx, testCoordinationNodePath, public.Resource{
+	err = db.Ratelimiter().AlterResource(ctx, testCoordinationNodePath, ratelimiter.Resource{
 		ResourcePath: testResource,
-		HierarchicalDrr: public.HierarchicalDrrSettings{
+		HierarchicalDrr: ratelimiter.HierarchicalDrrSettings{
 			MaxUnitsPerSecond:       3,
 			MaxBurstSizeCoefficient: 4,
 		},
@@ -132,26 +132,49 @@ func TestRatelimiter(t *testing.T) {
 		t.Fatal("Resource invalid")
 	}
 	// list resource
-	list, err := db.Ratelimiter().ListResource(ctx, testCoordinationNodePath, testResource, true)
+	list, err := db.Ratelimiter().ListResource(
+		ctx,
+		testCoordinationNodePath,
+		testResource,
+		true,
+	)
 	if err != nil {
 		t.Fatal(err)
 	}
 	if len(list) != 1 || list[0] != testResource {
 		t.Fatal("ListResource error")
 	}
-	// acquire resource amount <1,false>
+	// acquire resource amount 1
 	time.Sleep(time.Second) // for accumulate
-	err = db.Ratelimiter().AcquireResource(ctx, testCoordinationNodePath, testResource, 1, false)
+	err = db.Ratelimiter().AcquireResource(
+		ctx,
+		testCoordinationNodePath,
+		testResource,
+		1,
+		ratelimiter.WithAcquire(),
+	)
 	if err != nil {
 		t.Fatal(err)
 	}
-	// acquire resource amount <10000,true>
-	err = db.Ratelimiter().AcquireResource(ctx, testCoordinationNodePath, testResource, 10000, true)
+	// report resource amount 10000
+	err = db.Ratelimiter().AcquireResource(
+		ctx,
+		testCoordinationNodePath,
+		testResource,
+		10000,
+		ratelimiter.WithReportSync(),
+	)
 	if err != nil {
 		t.Fatal(err)
 	}
-	// acquire resource amount <10000,false>
-	err = db.Ratelimiter().AcquireResource(ctx, testCoordinationNodePath, testResource, 10000, false)
+	// acquire resource amount 10000
+	err = db.Ratelimiter().AcquireResource(
+		ctx,
+		testCoordinationNodePath,
+		testResource,
+		10000,
+		ratelimiter.WithAcquire(),
+	)
 	if err == nil {
 		t.Fatal("Resource must not be acquired")
 	}